requirements.txt
pandas
python-dotenv
requests
scikit-learn
matplotlib
plotly